Moved components that are part of other components to src/components
