+++ e-commerce-app/src/components/CartIcon.js
@@ -1,10 +1,10 @@
-import FilledCart from "./Icons/FilledCart.svg";
-import EmptyCart from "./Icons/EmptyCart.svg";
+import FilledCart from "../Icons/FilledCart.svg";
+import EmptyCart from "../Icons/EmptyCart.svg";
 
 import { useContext } from "react";
-import CartContext from "./CartContext";
+import CartContext from "../context/CartContext";
 
-import "./Styles/CartIcon.css";
+import "../Styles/CartIcon.css";
 
 const CartIconNavBar = () => {
   const { cart } = useContext(CartContext);
