+++ e-commerce-app/src/components/NavBar.js
@@ -1,7 +1,7 @@
 import { Link } from 'react-router-dom';
 import CartIconNavBar from './CartIcon';
 
-import './Styles/NavBar.css'
+import '../Styles/NavBar.css'
 
 //TODO: replace link text with images
