+++ e-commerce-app/src/components/Spinner.js
@@ -1,4 +1,4 @@
-import './Styles/Spinner.css';
+import '../Styles/Spinner.css';
 
 const Spinner = () => {
   return (
