+++ e-commerce-app/src/context/CartContext.js
@@ -1,8 +1,4 @@
 import React from "react";
-
-// const cartProduct = new Map();
-
-//cart structure sku:[product,quantity]
   
 const CartContext = React.createContext({
   cart: new Map(),
